requirements.txt
python-dotenv==1.0.0
mysql-connector-python==8.0.33
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6-sip==13.10.0
requests==2.31.0 